Add agency module and move button styling to global
This update consists of the introduction of a new agency module which includes components for agency creation, editing, and listing as well as corresponding routing. Additionally, the styling of the mat-button-wrapper has been moved from company-specific to global. This will provide consistent button styling throughout the application. Some updates to the company modules and error handling have also been included.